pyproject.toml
[project]
name = "anthropology-toolkit-service"
version = "0.1.0"
description = "Web service wrapping AI-powered qualitative analysis tools for anthropology research"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "python-multipart>=0.0.12",
    "openai>=1.50",
    "pydantic-settings>=2.5",
    "pdfplumber>=0.11",
    "python-docx>=1.1",
    "pandas>=2.2",
    "openpyxl>=3.1",
    "sentence-transformers>=3.3",
    "nltk>=3.9",
    "scikit-learn>=1.5",
    "numpy>=1.26",
    "matplotlib>=3.9",
    "networkx>=3.3",
]

[project.scripts]
serve = "toolkit.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/toolkit"]

[dependency-groups]
dev = [
    "mypy>=1.20.1",
    "pre-commit>=4.5.1",
    "ruff>=0.15.11",
]

[tool.ruff]
src = ["src"]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501", "B008", "SIM102"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
no_strict_optional = true
